+++ R/FishResp.R
@@ -2,8 +2,9 @@
 #'
 #' Calculates metabolic rate of fish and other aquatic organisms measured using an intermittent-flow respirometry approach. The tool is used to run a set of graphical QC tests of raw respirometry data, correct it for background respiration and chamber effect, filter and extract target values of absolute and mass-specific metabolic rate. Experimental design should include background respiration tests and measuring of one or more metabolic rate traits. The R package is ideally integrated with the pump controller 'PumpResp' and the DO meter 'SensResp' (open-source hardware by FishResp). Raw respirometry data can be also imported from 'AquaResp' (free software), 'AutoResp', ('LoligoSystems'), 'OxyView' ('PreSens'), 'Pyro Oxygen Logger' ('PyroScience'), and 'Q-box Aqua' ('QubitSystems').
 #'
-#' @docType package
+#' _PACKAGE
 #' @name FishResp
+#' @keywords internal
 #' @author Sergey Morozov \email{sergey.morozov@@helsinki.fi} Scott McCairns \email{scott.mccairns@@inra.fr}
 #'
 #' @references {