+++ R/convert.respirometry.R
@@ -12,10 +12,10 @@
 #' @param export.file  the name of a file with results of the DO unit conversion
 #' @param n.chamber  integer: the number of chambers used in an experiment (including empty ones)
 #' @param logger  string: the name of a logger software used for intermittent-flow respirometry. Note, that both 'OxyView' and 'Pyro Oxygen Logger' used in couple with the 'AquaResp' software should be converted to the 'FishResp' format before running this function (see the functions \code{\link{presens.aquaresp}} or \code{\link{pyroscience.aquaresp}}, respectively).
-#' @param from  string: dissolved oxygen unit in an imported file (more information can be found in the documentation of the function \code{\link{conv_o2}}, R package \pkg{respirometry})
-#' @param to  string: dissolved oxygen unit in an exported file (more information can be found in the documentation of the function \code{\link{conv_o2}}, R package \pkg{respirometry})
-#' @param sal  string: salinity is measured in ppm (more information can be found in the documentation of the function \code{\link{conv_o2}}, R package \pkg{respirometry})
-#' @param atm_pres  string: ambient atmospheric pressure value (more information can be found in the documentation of the function \code{\link{conv_o2}}, R package \pkg{respirometry})
+#' @param from  string: dissolved oxygen unit in an imported file (more information can be found in the documentation of the function \code{\link[respirometry]{conv_o2}}, R package \pkg{respirometry})
+#' @param to  string: dissolved oxygen unit in an exported file (more information can be found in the documentation of the function \code{\link[respirometry]{conv_o2}}, R package \pkg{respirometry})
+#' @param sal  string: salinity is measured in ppm (more information can be found in the documentation of the function \code{\link[respirometry]{conv_o2}}, R package \pkg{respirometry})
+#' @param atm_pres  string: ambient atmospheric pressure value (more information can be found in the documentation of the function \code{\link[respirometry]{conv_o2}}, R package \pkg{respirometry})
 #'
 #' @return The function exports a data frame with converted DO units.
 #'