+++ R/presens.aquaresp.R
@@ -1,6 +1,6 @@
 #' Convert Respirometry Data from PreSens and AquaResp Software to the FishResp Format
 #'
-#' The function is used to convert raw data from 'OxyView' (\href{https://www.presens.de}{PreSens}) and a summary file from 'AquaResp' (\href{https://www.aquaresp.com}{free software}) to 'FishResp' format. This function should be applied before usage of the functions \code{\link{import.test}} and \code{\link{import.meas}}. The output is a file containing raw respirometry data in the 'FishResp' format (see Details in \code{\link{import.test}} to read more information about the 'FishResp' format)
+#' The function is used to convert raw data from 'OxyView' (\href{https://www.presens.de}{PreSens}) and a summary file from 'AquaResp' (\href{http://bioold.science.ku.dk/jfsteffensen/AquaResp.htm}{free software}) to 'FishResp' format. This function should be applied before usage of the functions \code{\link{import.test}} and \code{\link{import.meas}}. The output is a file containing raw respirometry data in the 'FishResp' format (see Details in \code{\link{import.test}} to read more information about the 'FishResp' format)
 #'
 #' @usage
 #' presens.aquaresp(presens.file,
@@ -11,7 +11,7 @@
 #'                  wait.phase = NA, measure.phase = NA)
 #'
 #' @param presens.file  the name of a file which contains raw data obtained from the 'OxyView' software (\href{https://www.presens.de}{PreSens})
-#' @param aquaresp.file  the name of a file which contains summary data obtained from the 'AquaResp' software (\href{https://www.aquaresp.com}{free software})
+#' @param aquaresp.file  the name of a file which contains summary data obtained from the 'AquaResp' software (\href{http://bioold.science.ku.dk/jfsteffensen/AquaResp.htm}{free software})
 #' @param fishresp.file  the name of an exported file containing raw data in the 'FishResp' format
 #' @param n.chamber  integer: the number of chambers used in an experiment (including empty ones)
 #' @param date.format  string: date format (DMY, MDY or YMD) used in raw data obtained from the 'OxyView' software